Move highlighting from Contents to View
