+++ examples/highlights.html
@@ -43,6 +43,14 @@
       margin-top: 50px;
     }
 
+    [ref="epubjs-mk"] {
+      background: url("data:image/svg+xml;base64,PHN2ZyB2ZXJzaW9uPScxLjEnIHhtbG5zPSdodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZycgeG1sbnM6eGxpbms9J2h0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsnIHg9JzBweCcgeT0nMHB4JyB2aWV3Qm94PScwIDAgNzUgNzUnPjxnIGZpbGw9JyNCREJEQkQnIGlkPSdidWJibGUnPjxwYXRoIGNsYXNzPSdzdDAnIGQ9J00zNy41LDkuNEMxOS42LDkuNCw1LDIwLjUsNSwzNC4zYzAsNS45LDIuNywxMS4zLDcuMSwxNS42TDkuNiw2NS42bDE5LTcuM2MyLjgsMC42LDUuOCwwLjksOC45LDAuOSBDNTUuNSw1OS4yLDcwLDQ4LjEsNzAsMzQuM0M3MCwyMC41LDU1LjQsOS40LDM3LjUsOS40eicvPjwvZz48L3N2Zz4=") no-repeat;
+      width: 20px;
+      height: 20px;
+      cursor: pointer;
+      margin-left: 0;
+    }
+
   </style>
 </head>
 <body>
@@ -62,7 +70,8 @@
     var rendition = book.renderTo("viewer", {
       width: "100%",
       height: 600,
-      ignoreClass: 'annotator-hl'
+      ignoreClass: 'annotator-hl',
+      manager: "continuous"
     });
 
     var displayed = rendition.display(6);
