+++ package.json
@@ -74,7 +74,7 @@
     "event-emitter": "^0.3.5",
     "jszip": "^3.1.4",
     "lodash": "^4.17.4",
-    "marks-pane": "^1.0.2",
+    "marks-pane": "^1.0.4",
     "path-webpack": "0.0.3",
     "stream-browserify": "^2.0.1",
     "xmldom": "^0.1.27"
